+++ aslprep/cli/workflow.py
@@ -12,11 +12,11 @@ a hard-limited memory-scope.
 
 def build_workflow(config_file, retval):
     """Create the Nipype Workflow that supports the whole execution graph."""
+    from niworkflows.reports.core import generate_reports
     from niworkflows.utils.bids import check_pipeline_version, collect_participants
     from niworkflows.utils.misc import check_valid_fs_license
 
     from aslprep import config
-    from aslprep.niworkflows.reports import generate_reports
     from aslprep.utils.misc import check_deps
     from aslprep.workflows.base import init_aslprep_wf
