+++ aslprep/sdcflows/workflows/pepolar.py
@@ -18,7 +18,7 @@ from ...niworkflows.engine.workflows import LiterateWorkflow as Workflow
 from ...niworkflows.interfaces import CopyHeader
 from ...niworkflows.interfaces.freesurfer import StructuralReference
 from ...niworkflows.interfaces.registration import ANTSApplyTransformsRPT
-from ...niworkflows.func.util import init_enhance_and_skullstrip_asl_wf
+from aslprep.workflows.asl.util import init_enhance_and_skullstrip_asl_wf
 
 from nipype.pipeline import engine as pe
 from nipype.interfaces import afni, ants, utility as niu
