+++ aslprep/sdcflows/workflows/unwarp.py
@@ -4,8 +4,8 @@
 from nipype.pipeline import engine as pe
 from nipype.interfaces import utility as niu
 from niworkflows.engine.workflows import LiterateWorkflow as Workflow
-from aslprep.niworkflows.interfaces.registration import ANTSApplyTransformsRPT
-from ...niworkflows.func.util import init_enhance_and_skullstrip_asl_wf
+from niworkflows.interfaces.reportlets.registration import ANTSApplyTransformsRPT
+from aslprep.workflows.asl.util import init_enhance_and_skullstrip_asl_wf
 
 
 def init_sdc_unwarp_wf(omp_nthreads, debug, name='sdc_unwarp_wf'):
