+++ aslprep/utils/plotting.py
@@ -10,6 +10,7 @@ from nilearn import image, plotting
 from nilearn._utils.niimg import load_niimg
 from niworkflows import NIWORKFLOWS_LOG
 from niworkflows.interfaces.plotting import _get_tr
+from niworkflows.viz.plots import confoundplot, plot_carpet, spikesplot
 from niworkflows.viz.utils import (
     compose_view,
     cuts_from_bbox,
@@ -18,8 +19,6 @@ from niworkflows.viz.utils import (
 )
 from seaborn import color_palette
 from svgutils.transform import SVGFigure
-
-from aslprep.niworkflows.viz.plots import confoundplot, plot_carpet, spikesplot
 
 
 class ASLPlot:
