+++ docs/workflows.rst
@@ -65,7 +65,7 @@ brain extraction workflow:
    :graph2use: orig
    :simple_form: yes
 
-   from aslprep.niworkflows.anat.ants import init_brain_extraction_wf
+   from niworkflows.anat.ants import init_brain_extraction_wf
 
    wf = init_brain_extraction_wf()
 
@@ -143,7 +143,7 @@ This workflow estimates a reference image for an
 :abbr:`ASL (Arterial Spin Labelling)` series.
 The reference image is then used to calculate a brain mask for the
 :abbr:`ASL (Arterial Spin Labelling)` signal using *NiWorkflow's*
-:py:func:`~aslprep.niworkflows.func.util.init_enhance_and_skullstrip_asl_wf`.
+:py:func:`~aslprep.workflows.asl.util.init_enhance_and_skullstrip_asl_wf`.
 Subsequently, the reference image is fed to the :ref:`head-motion estimation
 workflow <asl_hmc>` and the :ref:`registration workflow <asl_reg>` to map the
 ASL series onto the T1w image of the same subject.
